LPS-179094 Extract ScreenNavigationEntry methods from ObjectDefinitionsLayoutsScreenNavigationCategory to ObjectDefinitionsLayoutsScreenNavigationEntry, implement getCategoryKey method in ObjectDefinitionsLayoutsScreenNavigationEntry and make ObjectDefinitionsLayoutsScreenNavigationCategory extends BaseObjectDefinitionsScreenNavigationCategory
